scripts/requirements-process-images.txt
# Requirements for process-raw-images-to-cdn.py script
boto3>=1.26.0
Pillow>=11.0.0
pillow-avif-plugin>=1.3.0
